pyproject.toml
[tool.poetry]
name = "confirm"
version = "0.1.0"
description = ""
authors = [
    "Confirm Solutions <research@confirmsol.org>",
]
packages = [
    { include = "confirm" },
    { include = "imprint", from = "imprint" },
]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
numpy = "^1.24.2"
scipy = "^1.10.1"
sympy = "^1.11.1"
matplotlib = "^3.7.0"
pandas = "^1.5.3"
jax = "^0.4.4"
numpyro = "^0.11.0"
duckdb = "^0.7.1"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"
pre-commit = "^3.1.1"
pytest-xdist = {extras = ["psutil"], version = "^3.2.0"}
ipython = "^8.11.0"
nbformat = "^5.7.3"

[tool.poetry.group.cloud]
optional = true

[tool.poetry.group.cloud.dependencies]
pyarrow = "^10.0.1"
boto3 = "^1.26.81"
cloudpickle = [
    { version = ">=2.0.0,<2.1.0", python = "<3.11" },
    { version = ">=2.2.0,<2.3.0", python = ">=3.11" },
]
clickhouse-connect = "^0.5.13"
python-dotenv = "^0.21.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.1.0"}
line-profiler = "^4.0.2"
jupytext = "^1.14.5"
reorder-python-imports = "^3.9.0"
flake8 = "^6.0.0"
typer = "^0.7.0"
jupyter = "^1.0.0"
rich = "^13.3.1"
ipdb = "^0.13.11"
tomli-w = "^1.0.0"

[tool.poetry.group.cloud_dev]
optional = true

[tool.poetry.group.cloud_dev.dependencies]
aws-cdk-lib = "^2.66.1"
"aws-cdk.aws-batch-alpha" = "^2"
modal-client = "^0.46.1403"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-p imprint.testing --capture=no --show-capture=no --tb=short --strict-markers --doctest-modules"
testpaths = ["tests", "imprint/imprint"]

[tool.jupytext]
notebook_metadata_filter = "-all"
cell_metadata_filter = "-all"

[build-system]
requires = [
    "poetry-core>=1.0.0",
]
build-backend = "poetry.core.masonry.api"
